controllers

+++ app/Http/Controllers/AuthorController.php
@@ -8,19 +8,22 @@ use App\Http\Requests\CreateAuthorRequest;
 
 class AuthorController extends Controller
 {
-    public function index(){
+    public function index()
+    {
         // $data = Author::withTrashed()   ->orderby('id','ASC')->get();
-        $data= Author::paginate(10);
+        $data = Author::paginate(10);
 
         // $data= Author::all();
         return view('authors.index', compact('data'));
     }
 
-    public function create(){
+    public function create()
+    {
         return view('authors.create');
     }
 
-    public function store(CreateAuthorRequest $request){
+    public function store(CreateAuthorRequest $request)
+    {
 
         // $validated = $request->validate([
         //     'name' => 'required|max:255',
@@ -37,6 +40,11 @@ class AuthorController extends Controller
         //     ];
         // Author::create($dataToInsert);
 
+        $counter = Author::where('name', '=', $request->name)->count();
+        if ($counter > 0) {
+            return redirect()->back()->with(['error' => 'Author already exists',])->withInput();
+        }
+
         $author = new Author();
         $author->name = $request->name;
         $author->bio = $request->bio;
@@ -44,13 +52,18 @@ class AuthorController extends Controller
         return redirect()->route('authors.index');
     }
 
-    public function edit($id){
+    public function edit($id)
+    {
         $author = Author::find($id);
         return view('authors.update', compact('author'));
     }
 
-    public function update(CreateAuthorRequest $request, $id){
+    public function update(CreateAuthorRequest $request, $id)
+    {
         $author = Author::find($id);
+        if (empty($author)) {
+            return redirect()->back()->with(['error' => 'Author not found',])->withInput();
+        }
         $author->name = $request->name;
         $author->bio = $request->bio;
         $author->save();
@@ -58,19 +71,20 @@ class AuthorController extends Controller
         // $DataToUpdate['name'] = $request->name;
         // $DataToUpdate['bio'] = $request->bio;
         // Author::where("id", "=", $id)->update($DataToUpdate);
-        
+
         return redirect()->route('authors.index');
     }
 
-    public function delete($id){
+    public function delete($id)
+    {
         $author = Author::find($id);
+        if (empty($author)){
+            return redirect()->back()->with(['error' => 'Author not found',])->withInput();
+        }
         $author->delete();
 
         // Author::where('id', '=', $id)->forceDelete();
 
         return redirect()->route('authors.index');
     }
-
-
 }
-
